Improve the folder structure of the app
