+++ src/apps/admin/layout/AppBar.jsx
@@ -8,7 +8,7 @@ import {
   Button,
 } from '@material-ui/core';
 import MenuIcon from '@material-ui/icons/Menu';
-import DropDown from '../../Layout/DropDownMenu';
+import DropDown from '../../common/DropDownMenu';
 import { makeStyles } from '@material-ui/core/styles';
 import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
