+++ src/apps/common/Register.jsx
@@ -2,16 +2,16 @@ import React, { Component } from 'react';
 import { Redirect, Link } from 'react-router-dom';
 import { Paper, TextField, Button, Typography } from '@material-ui/core';
 import { connect } from 'react-redux';
-import { login } from './../store/slices/auth/actions';
-import StyledFirebaseAuth from 'react-firebaseui/StyledFirebaseAuth';
-import firebase, { uiConfig } from './../config/firebase-config';
+import { register } from './../../store/slices/auth/actions';
 
-const COMPONENT = 'Login';
+const COMPONENT = 'Register';
 
-class Login extends Component {
+class Register extends Component {
   state = {
     email: '',
     password: '',
+    firstName: '',
+    lastName: '',
     shouldRedirect: false,
   };
 
@@ -21,10 +21,12 @@ class Login extends Component {
     });
   };
 
-  handleLogin = async (e) => {
+  handleSignup = async (e) => {
     e.preventDefault();
     try {
-      await this.props.login({
+      await this.props.register({
+        firstName: this.state.firstName,
+        lastName: this.state.lastName,
         email: this.state.email,
         password: this.state.password,
       });
@@ -43,7 +45,7 @@ class Login extends Component {
 
     return (
       <Paper elevation={3}>
-        <h1>Login</h1>
+        <h1>Register</h1>
         <form>
           <TextField
             required
@@ -57,6 +59,26 @@ class Login extends Component {
           />
           <TextField
             required
+            id='firstName'
+            name='firstName'
+            label='Firstame'
+            variant='outlined'
+            placeholder='Firstname'
+            onChange={this.handleChange}
+            value={this.state.firstName}
+          />
+          <TextField
+            required
+            id='lastName'
+            name='lastName'
+            label='Lastname'
+            variant='outlined'
+            placeholder='Lastname'
+            onChange={this.handleChange}
+            value={this.state.lastName}
+          />
+          <TextField
+            required
             id='password'
             name='password'
             label='Password'
@@ -67,20 +89,16 @@ class Login extends Component {
             value={this.state.password}
           />
           <Button
-            onClick={this.handleLogin}
             variant='contained'
             color='primary'
+            onClick={this.handleSignup}
           >
-            Login
+            Signup
           </Button>
         </form>
-        <Typography component={Link} to='/register'>
-          Create a new account!
+        <Typography component={Link} to='/login'>
+          Already have an account? Login!
         </Typography>
-        <StyledFirebaseAuth
-          uiConfig={uiConfig}
-          firebaseAuth={firebase.auth()}
-        />
         {authError && <Typography>{authError}</Typography>}
       </Paper>
     );
@@ -94,8 +112,10 @@ const mapStateToProps = (state) => {
   };
 };
 
-const mapDispatchToProps = (dispatch) => ({
-  login: (credentials) => dispatch(login(credentials)),
-});
+const mapDispatchToProps = (dispatch) => {
+  return {
+    register: (credentials) => dispatch(register(credentials)),
+  };
+};
 
-export default connect(mapStateToProps, mapDispatchToProps)(Login);
+export default connect(mapStateToProps, mapDispatchToProps)(Register);
